+++ robot_descriptions/gen3_lite_description.py
@@ -2,9 +2,9 @@
 # -*- coding: utf-8 -*-
 #
 # SPDX-License-Identifier: Apache-2.0
-# Copyright 2023 Inria
+# Copyright 2025 Inria
 
-"""ergoCub description."""
+"""Kinova Gen3 Lite description."""
 
 from os import getenv as _getenv
 from os import path as _path
@@ -12,12 +12,10 @@ from os import path as _path
 from ._cache import clone_to_cache as _clone_to_cache
 
 REPOSITORY_PATH: str = _clone_to_cache(
-    "ergocub-software",
+    "ros2_kortex",
     commit=_getenv("ROBOT_DESCRIPTION_COMMIT", None),
 )
 
-PACKAGE_PATH: str = _path.join(REPOSITORY_PATH, "urdf", "ergoCub")
+PACKAGE_PATH: str = _path.join(REPOSITORY_PATH, "kortex_description")
 
-URDF_PATH: str = _path.join(
-    PACKAGE_PATH, "robots", "ergoCubSN002", "model.urdf"
-)
+URDF_PATH: str = _path.join(PACKAGE_PATH, "robots", "gen3_lite.urdf")